Extract string chunking utils
